requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
opencv-contrib-python
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
jupyter	
PIMS @ git+https://github.com/soft-matter/pims.git#egg=PIMS	
av	
fastapi	
uvicorn	
streamlit	
protobuf	
python-multipart

# logging -------------------------------------
tensorboard>=2.4.1
wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization

# extras -------------------------------------
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
thop  # FLOPs computation

#tutorial extras
ipyfilechooser
gdown
itables
#getpass
panoptes-client
paramiko
scp